+++ templates/dashboard.html
@@ -22,7 +22,7 @@
         <a class="nav-link active" href="/statistics">General Channel Statistics</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="/custom_statistics">Custom Statistics</a>
+        <a class="nav-link" href="/video_statistics">Video Statistics</a>
       </li>
     </ul>
     <div style="margin: 1%;">
